refactor(card)#34: CardApprovalItemResponse approvalAmount 타입 BigDecimal로 변경
- Long → BigDecimal: DB 스키마 DECIMAL(18,2) 기준에 맞게 수정
- 금융 금액 역직렬화 시 정밀도 손실 방지

+++ src/main/java/com/mydata/domain/card/client/dto/response/CardApprovalItemResponse.java
@@ -1,5 +1,6 @@
 package com.mydata.domain.card.client.dto.response;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -15,7 +16,7 @@ public class CardApprovalItemResponse {
 
   private String merchantCategory;
 
-  private Long approvalAmount;
+  private BigDecimal approvalAmount;
 
   private String approvalStatus;
 }
